Gather rustc-specific functions around `MatchCheckCtxt`
